Change f64 dtypes in annotations to f32.

+++ src/annotations/bounding_box.rs
@@ -3,8 +3,8 @@ use std::fmt;
 /// A set of custom errors for more informative error handling.
 #[derive(Debug)]
 pub enum BoundingBoxError {
-    InvalidLeftRight { left: f64, right: f64 },
-    InvalidTopBottom { top: f64, bottom: f64 },
+    InvalidLeftRight { left: f32, right: f32 },
+    InvalidTopBottom { top: f32, bottom: f32 },
 }
 
 impl fmt::Display for BoundingBoxError {
@@ -42,20 +42,20 @@ impl std::error::Error for BoundingBoxError {}
 /// of the image being y=0.
 #[derive(Debug)]
 pub struct BoundingBox {
-    left: f64,
-    top: f64,
-    right: f64,
-    bottom: f64,
+    left: f32,
+    top: f32,
+    right: f32,
+    bottom: f32,
     category: String,
 }
 
 impl BoundingBox {
     /// Checks if a box has valid parameters before constructing.
     pub fn new(
-        left: f64,
-        top: f64,
-        right: f64,
-        bottom: f64,
+        left: f32,
+        top: f32,
+        right: f32,
+        bottom: f32,
         category: String,
     ) -> Result<Self, BoundingBoxError> {
         if left > right {
@@ -89,30 +89,30 @@ impl fmt::Display for BoundingBox {
 /// Any annotation that uses a bounding box as its base has "box like geometry", and therefore can
 /// be passed in to useful functions like non-maximum suppression and intersection over union.
 pub trait BoundingBoxGeometry {
-    fn left(&self) -> f64;
-    fn top(&self) -> f64;
-    fn right(&self) -> f64;
-    fn bottom(&self) -> f64;
+    fn left(&self) -> f32;
+    fn top(&self) -> f32;
+    fn right(&self) -> f32;
+    fn bottom(&self) -> f32;
     fn category(&self) -> &str;
-    fn area(&self) -> f64;
-    fn center(&self) -> (f64, f64);
-    fn as_xyxy(&self) -> (f64, f64, f64, f64);
+    fn area(&self) -> f32;
+    fn center(&self) -> (f32, f32);
+    fn as_xyxy(&self) -> (f32, f32, f32, f32);
 }
 
 impl BoundingBoxGeometry for BoundingBox {
-    fn left(&self) -> f64 {
+    fn left(&self) -> f32 {
         self.left
     }
 
-    fn top(&self) -> f64 {
+    fn top(&self) -> f32 {
         self.top
     }
 
-    fn right(&self) -> f64 {
+    fn right(&self) -> f32 {
         self.right
     }
 
-    fn bottom(&self) -> f64 {
+    fn bottom(&self) -> f32 {
         self.bottom
     }
 
@@ -120,18 +120,18 @@ impl BoundingBoxGeometry for BoundingBox {
         &self.category
     }
 
-    fn area(&self) -> f64 {
+    fn area(&self) -> f32 {
         (self.right - self.left) * (self.bottom - self.top)
     }
 
-    fn center(&self) -> (f64, f64) {
+    fn center(&self) -> (f32, f32) {
         (
-            0.5_f64 * (self.right - self.left),
-            0.5_f64 * (self.bottom - self.top),
+            0.5_f32 * (self.right - self.left),
+            0.5_f32 * (self.bottom - self.top),
         )
     }
 
-    fn as_xyxy(&self) -> (f64, f64, f64, f64) {
+    fn as_xyxy(&self) -> (f32, f32, f32, f32) {
         (self.left, self.top, self.right, self.bottom)
     }
 }
